requirements.txt
gradio
openai
numpy
flask
uvicorn
gunicorn
python-dotenv